requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.3
django-filter==21.1
django-froala-editor==4.0.10
django-heroku==0.3.1
djangorestframework==3.13.1
djongo==1.3.6
gunicorn==20.1.0
importlib-metadata==4.11.3
Markdown==3.3.6
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pymongo==4.0.2
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.2.4
whitenoise==6.0.0
zipp==3.7.0
